kinda works so far

+++ CaesarAIMovieStream/CaesarAIRealDebrid/responses/StatusAndProgressResponse.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 class StatusAndProgressResponse(BaseModel):
     id:str
     status:str
-    progress:int
+    progress:float
